modified:   CMakeLists.txt 	deleted:    audi_a8_adapter.launch 	modified:   include/audi_a8_can/CANTrans.h 	modified:   include/audi_a8_can/can_translator.h 	deleted:    package.xml 	renamed:    parameters/audi_a8/Flexray_CAN.dbc -> parameters/Flexray_CAN.dbc 	deleted:    parameters/audi_a8_can_params.yaml 	renamed:    parameters/audi_a8/key.map -> parameters/key.map 	renamed:    parameters/audi_a8/trajectory_draft.dbc -> parameters/trajectory_draft.dbc 	renamed:    parameters/audi_a8/unit.map -> parameters/unit.map 	modified:   src/audi_a8_control.cpp 	modified:   src/audi_a8_control.h 	modified:   src/audi_a8_msgs_processor.cpp 	modified:   src/audi_a8_msgs_processor.h

+++ src/can_matrix_spi/spi_dbus_hal.cpp
@@ -355,14 +355,14 @@ error:
     return -1;
 }
 
-
-float GetTimeStamp()//
+double GetTimeStamp()//
 {
-float timestamp;
+double timestamp;
 struct timeval tv;
 gettimeofday(&tv,NULL); //gettimeofday(&start,&tz);结果一样
-timestamp = tv.tv_sec+tv.tv_usec/1000000;
+timestamp = double(tv.tv_sec)+double(tv.tv_usec)/1000000;
 return timestamp;
 }
+
